requirements.txt
python-telegram-bot>=20.0
sqlalchemy>=2.0.0
aiohttp>=3.8.0
psutil>=5.9.0
python-dateutil>=2.8.2
marzpy>=1.0.0
cryptography>=40.0.0
aiosqlite>=0.19.0 